+++ api/kvstore.proto
@@ -1,0 +1,60 @@
+syntax = "proto3";
+
+package kvstore;
+
+service KVStore {
+    rpc Put(PutRequest) returns (PutResponse);
+    rpc Get(GetRequest) returns (GetResponse);
+    rpc Delete(DeleteRequest) returns (DeleteResponse);
+    rpc Status(StatusRequest) returns (StatusResponse);
+    rpc Join(JoinRequest) returns (JoinResponse);
+    rpc Leave(LeaveRequest) returns (LeaveResponse);
+}
+
+message PutRequest {
+    string key = 1;
+    string value = 2;
+}
+message PutResponse {
+    bool success = 1;
+    string error = 2;
+}
+
+message GetRequest {
+    string key = 1;
+}
+message GetResponse {
+    bool found = 1;
+    string value = 2;
+    string error = 3;
+}
+
+message DeleteRequest {
+    string key = 1;
+}
+message DeleteResponse {
+    bool success = 1;
+    string error = 2;
+}
+
+message StatusRequest {}
+message StatusResponse {
+    string role = 1;     // "Leader" or "Follower"
+    int32 node_id = 2;
+    int32 term = 3;
+    string leader_address = 4;
+}
+
+message JoinRequest {
+    string address = 1; // 新ノードのgRPCアドレス
+}
+message JoinResponse {
+    bool success = 1;
+    string error = 2;
+}
+
+message LeaveRequest {}
+message LeaveResponse {
+    bool success = 1;
+    string error = 2;
+}
