+++ src/client.cpp
@@ -1,53 +1,69 @@
+#include "kvstore.grpc.pb.h"
+#include <grpcpp/grpcpp.h>
 #include <iostream>
-#include <memory>
 #include <string>
 
-#include "hello.grpc.pb.h"
-#include <grpcpp/grpcpp.h>
+int main(int argc, char **argv) {
+    std::string address = "localhost:12345";
+    auto channel = grpc::CreateChannel(address, grpc::InsecureChannelCredentials());
+    auto stub = kvstore::KVStore::NewStub(channel);
 
-using grpc::Channel;
-using grpc::ClientContext;
-using grpc::Status;
-using helloworld::HelloReply;
-using helloworld::HelloRequest;
-using helloworld::HelloService;
-
-class HelloClient {
-public:
-    HelloClient(std::shared_ptr<Channel> channel)
-        : stub_(HelloService::NewStub(channel)) {}
-
-    std::string SayHello(const std::string &user) {
-        HelloRequest request;
-        request.set_name(user);
-
-        HelloReply reply;
-        ClientContext context;
-
-        Status status = stub_->SayHello(&context, request, &reply);
-
-        if (status.ok()) {
-            return reply.message();
+    // --- Put ---
+    {
+        kvstore::PutRequest req;
+        kvstore::PutResponse res;
+        req.set_key("hello");
+        req.set_value("world");
+        grpc::ClientContext ctx;
+        auto status = stub->Put(&ctx, req, &res);
+        if (status.ok() && res.success()) {
+            std::cout << "[Put] OK" << std::endl;
         } else {
-            return "RPC failed";
+            std::cout << "[Put] Failed: " << res.error() << std::endl;
         }
     }
 
-private:
-    std::unique_ptr<HelloService::Stub> stub_;
-};
-
-int main(int argc, char **argv) {
-    HelloClient client(grpc::CreateChannel(
-        "localhost:50051", grpc::InsecureChannelCredentials()));
-
-    std::string user("world");
-    if (argc > 1) {
-        user = argv[1];
+    // --- Get ---
+    {
+        kvstore::GetRequest req;
+        kvstore::GetResponse res;
+        req.set_key("hello");
+        grpc::ClientContext ctx;
+        auto status = stub->Get(&ctx, req, &res);
+        if (status.ok() && res.found()) {
+            std::cout << "[Get] " << req.key() << " → " << res.value() << std::endl;
+        } else {
+            std::cout << "[Get] Not found or error: " << res.error() << std::endl;
+        }
     }
 
-    std::string reply = client.SayHello(user);
-    std::cout << "Received: " << reply << std::endl;
+    // --- Delete ---
+    {
+        kvstore::DeleteRequest req;
+        kvstore::DeleteResponse res;
+        req.set_key("hello");
+        grpc::ClientContext ctx;
+        auto status = stub->Delete(&ctx, req, &res);
+        if (status.ok() && res.success()) {
+            std::cout << "[Delete] OK" << std::endl;
+        } else {
+            std::cout << "[Delete] Failed: " << res.error() << std::endl;
+        }
+    }
+
+    // --- Get again ---
+    {
+        kvstore::GetRequest req;
+        kvstore::GetResponse res;
+        req.set_key("hello");
+        grpc::ClientContext ctx;
+        auto status = stub->Get(&ctx, req, &res);
+        if (status.ok() && res.found()) {
+            std::cout << "[Get] " << req.key() << " → " << res.value() << std::endl;
+        } else {
+            std::cout << "[Get after delete] Not found or error: " << res.error() << std::endl;
+        }
+    }
 
     return 0;
 }
